requirements.txt
Brotli==1.0.9
inflate64==0.3.1
multivolumefile==0.2.3
pefile==2023.2.7
psutil==5.9.5
py7zr==0.20.5
pybcj==1.0.1
pycryptodomex==3.18.0
pyppmd==1.0.0
python-magic==0.4.27
pyzstd==0.15.9
rarfile==4.0
texttable==1.6.7
